gpio/pca953x: increase variables size to support 24 bit of data
Increase variable size from u16 to u32 to allocate 24 bit
of data required for the TCA6424 I/O expander device type

+++ drivers/gpio/gpio-pca953x.c
@@ -78,10 +78,10 @@ struct pca953x_chip {
 
 #ifdef CONFIG_GPIO_PCA953X_IRQ
 	struct mutex irq_lock;
-	uint16_t irq_mask;
-	uint16_t irq_stat;
-	uint16_t irq_trig_raise;
-	uint16_t irq_trig_fall;
+	uint32_t irq_mask;
+	uint32_t irq_stat;
+	uint32_t irq_trig_raise;
+	uint32_t irq_trig_fall;
 	int	 irq_base;
 #endif
 
@@ -349,8 +349,8 @@ static void pca953x_irq_bus_lock(struct irq_data *d)
 static void pca953x_irq_bus_sync_unlock(struct irq_data *d)
 {
 	struct pca953x_chip *chip = irq_data_get_irq_chip_data(d);
-	uint16_t new_irqs;
-	uint16_t level;
+	uint32_t new_irqs;
+	uint32_t level;
 
 	/* Look for any newly setup interrupt */
 	new_irqs = chip->irq_trig_fall | chip->irq_trig_raise;
@@ -368,8 +368,8 @@ static void pca953x_irq_bus_sync_unlock(struct irq_data *d)
 static int pca953x_irq_set_type(struct irq_data *d, unsigned int type)
 {
 	struct pca953x_chip *chip = irq_data_get_irq_chip_data(d);
-	uint16_t level = d->irq - chip->irq_base;
-	uint16_t mask = 1 << level;
+	uint32_t level = d->irq - chip->irq_base;
+	uint32_t mask = 1 << level;
 
 	if (!(type & IRQ_TYPE_EDGE_BOTH)) {
 		dev_err(&chip->client->dev, "irq %d: unsupported type %d\n",
@@ -399,12 +399,12 @@ static struct irq_chip pca953x_irq_chip = {
 	.irq_set_type		= pca953x_irq_set_type,
 };
 
-static uint16_t pca953x_irq_pending(struct pca953x_chip *chip)
+static uint32_t pca953x_irq_pending(struct pca953x_chip *chip)
 {
 	u32 cur_stat;
-	uint16_t old_stat;
-	uint16_t pending;
-	uint16_t trigger;
+	uint32_t old_stat;
+	uint32_t pending;
+	uint32_t trigger;
 	int ret, offset = 0;
 
 	switch (chip->chip_type) {
@@ -421,7 +421,6 @@ static uint16_t pca953x_irq_pending(struct pca953x_chip *chip)
 
 	/* Remove output pins from the equation */
 	cur_stat &= chip->reg_direction;
-
 	old_stat = chip->irq_stat;
 	trigger = (cur_stat ^ old_stat) & chip->irq_mask;
 
@@ -440,8 +439,8 @@ static uint16_t pca953x_irq_pending(struct pca953x_chip *chip)
 static irqreturn_t pca953x_irq_handler(int irq, void *devid)
 {
 	struct pca953x_chip *chip = devid;
-	uint16_t pending;
-	uint16_t level;
+	uint32_t pending;
+	uint32_t level;
 
 	pending = pca953x_irq_pending(chip);
